requirements.txt
cloudpickle==2.2.1
fastapi==0.92.0
uvicorn==0.21.0
mlflow-skinny==2.2.1
scikit-learn==1.2.1
pandas==1.5.3
tpot==0.11.7
numpy==1.23.5
scipy==1.10.1
